Port the parser tests
